[refactor] ThreadListHeader fetching 로직을 ThreadListBox로 분리
