+++ client/src/components/ThreadListBox/ThreadListHeader/ThreadListHeader.tsx
@@ -1,15 +1,12 @@
-import React, { useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
-import { useChannelState, useUserState } from '@/hooks';
+import React, { useState } from 'react';
 import styled from 'styled-components';
+import { useChannelState } from '@/hooks';
 import { CHANNEL_TYPE } from '@/utils/constants';
 import { JoinedUser } from '@/types';
 import { Link, useParams } from 'react-router-dom';
-import { loadChannelRequest } from '@/store/modules/channel.slice';
 import { DimModal, LockIcon, PoundIcon, WarningIcon, AddUserIcon } from '@/components';
 import theme from '@/styles/theme';
 import { flex, hoverActive } from '@/styles/mixin';
-import { enterRoomRequest, leaveRoomRequest } from '@/store/modules/socket.slice';
 import { AddUsersModalHeader, AddUsersModalBody } from './ChannelModal/AddUsersModal';
 import { AddTopicModalHeader, AddTopicModalBody } from './ChannelModal/AddTopicModal';
 import { ShowUsersModalHeader, ShowUsersModalBody } from './ChannelModal/ShowUsersModal';
@@ -95,30 +92,12 @@ interface RightSideParams {
 }
 
 const ThreadListHeader = () => {
-  const dispatch = useDispatch();
   const { current, users } = useChannelState();
-  const { userInfo } = useUserState();
+
   const [addUsersModalVisible, setAddUsersModalVisible] = useState(false);
   const [addTopicModalVisible, setAddTopicModalVisible] = useState(false);
   const [showUsersModalVisible, setShowUsersModalVisible] = useState(false);
   const { channelId }: RightSideParams = useParams();
-
-  useEffect(() => {
-    if (userInfo) {
-      dispatch(loadChannelRequest({ channelId: +channelId, userId: userInfo.id }));
-    }
-  }, [channelId, userInfo]);
-
-  useEffect(() => {
-    if (current) {
-      dispatch(enterRoomRequest({ room: current.name }));
-    }
-    return () => {
-      if (current) {
-        dispatch(leaveRoomRequest({ room: current.name }));
-      }
-    };
-  }, [current]);
 
   const clickShowUsersModal = () => {
     setShowUsersModalVisible((state) => !state);
